requirements.txt

bz2file==0.98

dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1

joblib==1.2.0

matplotlib==3.6.0
numpy==1.23.3

pandas==1.5.0
Pillow==9.2.0
plotly==5.10.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
sklearn==0.0
streamlit==1.13.0

threadpoolctl==3.1.0


watchdog==2.1.9


xgboost==1.6.2

